review: split progression.rs and address grant_cash test feedback
Three Copilot suggestions for the live-DB grant_cash tests:

1. File at 511 lines was over the repo's 500-line soft cap — the natural
   seam between the handler code and the test module makes the split
   uncontroversial. Promote progression.rs to a directory module
   (mod.rs + tests.rs) following the world_data/, outbox/, and move_/
   precedents. mod.rs ends at 353 lines, tests.rs at 165.

2. Rename `drops_grant_silently_when_player_row_missing` to
   `does_not_credit_when_player_row_missing`. The function emits
   `tracing::warn!` on the missing-row path, so "silently" was
   misleading — same correction the merged #139 review applied to
   send_store_open's bail-path tests.

3. Tighten the doc comment on the missing-row test. The previous wording
   claimed "must not emit a wire packet", but the test sets `connected`
   to an empty map, so any wire emit is skipped regardless of what the
   handler did. Narrow the comment to the DB invariants the test
   actually pins (no panic, no INSERT, bystander untouched).
